Reduce tenant API
